refactor(client): split crypto_service into init, peer-identity, otp parts (#761)
