+++ apps/client/lib/src/services/crypto_service.dart
@@ -19,6 +19,7 @@ import 'package:flutter/foundation.dart';
 import 'package:http/http.dart' as http;
 import 'package:shared_preferences/shared_preferences.dart';
 
+import 'crypto_exceptions.dart';
 import 'debug_log_service.dart';
 import 'safety_number_service.dart';
 import 'secure_key_store.dart';
@@ -26,78 +27,14 @@ import 'session_cache.dart';
 import 'signal_session.dart';
 import 'signal_x3dh.dart';
 
-/// Thrown by [CryptoService.getOrCreateSession] when the peer's identity
-/// key on the server differs from the key we previously trusted (TOFU
-/// violation). Callers must surface this via the safety-number /
-/// "trust new key" UI before establishing a new session — silently
-/// overwriting the trusted key would let an attacker who has taken over
-/// the server impersonate the peer (#580).
-class IdentityKeyChangedException implements Exception {
-  /// Peer whose identity key changed.
-  final String peerUserId;
+// Re-export the typed exceptions so external callers that already import
+// them from `crypto_service.dart` (e.g. websocket_provider, the test suite)
+// keep compiling without a coordinated import-rewrite across the codebase.
+export 'crypto_exceptions.dart';
 
-  /// Previously-trusted identity key (base64), or `null` if for some
-  /// reason the old value could not be read back.
-  final String? oldIdentityKeyB64;
-
-  /// New identity key returned by the server (base64).
-  final String newIdentityKeyB64;
-
-  const IdentityKeyChangedException({
-    required this.peerUserId,
-    required this.oldIdentityKeyB64,
-    required this.newIdentityKeyB64,
-  });
-
-  @override
-  String toString() =>
-      'IdentityKeyChangedException(peer=$peerUserId): trusted identity key '
-      'changed; user must verify safety number or explicitly accept new key';
-}
-
-/// Thrown by [CryptoService.uploadKeys] when the server rejects the upload
-/// because the identity-key fingerprint we presented for this device does
-/// not match what is on file (#664). The body carries the device_id and the
-/// expected/actual fingerprints (base64) so the UI can drive a typed reset
-/// flow instead of parsing English error strings.
-class IdentityKeyConflictException implements Exception {
-  /// Device id reported by the server. Falls back to the local device id when
-  /// the server returned a legacy (string-only) 409 without a JSON body.
-  final int deviceId;
-
-  /// Server-side fingerprint (base64) -- the value we'd need to match. Null
-  /// when the server returned a legacy body.
-  final String? expectedFingerprint;
-
-  /// Fingerprint we sent (base64). Null when no body was returned.
-  final String? actualFingerprint;
-
-  const IdentityKeyConflictException({
-    required this.deviceId,
-    this.expectedFingerprint,
-    this.actualFingerprint,
-  });
-
-  @override
-  String toString() =>
-      'IdentityKeyConflictException(device=$deviceId, '
-      'expected=${expectedFingerprint ?? "?"}, '
-      'actual=${actualFingerprint ?? "?"})';
-}
-
-/// Thrown by [CryptoService.decryptMessage] when an initial X3DH wire
-/// failed to authenticate (AES-GCM tag check). The receiver's keys on the
-/// server are likely stale; the service schedules an `uploadKeys()` to heal
-/// the bundle so the next message from the same peer can decrypt (#662).
-class InitialDecryptFailedException implements Exception {
-  final String peerUserId;
-  const InitialDecryptFailedException(this.peerUserId);
-
-  @override
-  String toString() =>
-      'InitialDecryptFailedException(peer=$peerUserId): initial X3DH wire '
-      'failed AES-GCM auth; uploaded fresh keys to heal stale bundle';
-}
+part 'crypto/init_extension.dart';
+part 'crypto/peer_identity_extension.dart';
+part 'crypto/otp_extension.dart';
 
 class CryptoService {
   static const _deviceIdPref = 'echo_device_id';
@@ -205,463 +142,11 @@ class CryptoService {
   /// This means old encrypted messages cannot be decrypted.
   bool get keysWereRegenerated => _keysWereRegenerated;
 
-  /// Migrate crypto keys from SharedPreferences to SecureKeyStore.
-  ///
-  /// Checks SharedPreferences for each crypto key; if found, copies it to
-  /// secure storage and deletes from SharedPreferences. Session keys (prefixed
-  /// with [_sessionPrefix]), OTP private keys, and peer identity keys are also
-  /// migrated.
-  ///
-  /// A completion flag is written to secure storage after all keys are
-  /// successfully moved. If the flag already exists, migration is skipped.
-  /// Each write is guarded: if [SecureKeyStore.write] throws (e.g. keyring
-  /// unavailable), the key is left in SharedPreferences so the next launch can
-  /// retry the migration instead of losing the data.
-  Future<void> _migrateFromSharedPreferences() async {
-    final store = SecureKeyStore.instance;
+  // Init / migration / signed-prekey rotation helpers live in
+  // crypto/init_extension.dart (extension CryptoServiceInit).
 
-    // Skip if migration was already completed.
-    final migrated = await store.read('_crypto_migration_complete');
-    if (migrated == 'true') return;
-
-    final prefs = await SharedPreferences.getInstance();
-    var allSucceeded = true;
-
-    // On web, skip removing keys from SharedPreferences after copying to
-    // secure storage. SecureKeyStore on web uses Web Crypto API encryption
-    // which can fail to decrypt after page refresh in some browsers, so
-    // SharedPreferences (plain localStorage) is the reliable fallback.
-    final removeFromPrefs = !kIsWeb;
-
-    // Migrate identity + signing + signed prekey (named keys)
-    for (final key in _allCryptoKeys) {
-      final value = prefs.getString(key);
-      if (value != null) {
-        try {
-          await store.write(key, value);
-          if (removeFromPrefs) await prefs.remove(key);
-          debugPrint(
-            '[Crypto] Migrated $key from SharedPreferences to '
-            'secure storage',
-          );
-        } catch (e) {
-          allSucceeded = false;
-          debugPrint(
-            '[Crypto] Migration of $key failed (keeping in SharedPreferences '
-            'for next attempt): $e',
-          );
-        }
-      }
-    }
-
-    // Migrate prefixed keys: sessions, OTP privates, peer identities
-    for (final key in prefs.getKeys()) {
-      final isPrefixed =
-          key.startsWith(_sessionPrefix) ||
-          key.startsWith(_otpPrivatePrefix) ||
-          key.startsWith(_peerIdentityPrefix) ||
-          key.startsWith(_peerIdentityChangedPrefix);
-      if (isPrefixed) {
-        final value = prefs.getString(key);
-        if (value != null) {
-          try {
-            await store.write(key, value);
-            if (removeFromPrefs) await prefs.remove(key);
-            debugPrint('[Crypto] Migrated $key to secure storage');
-          } catch (e) {
-            allSucceeded = false;
-            debugPrint(
-              '[Crypto] Migration of $key failed '
-              '(keeping in SharedPreferences): $e',
-            );
-          }
-        }
-      }
-    }
-
-    // Also migrate device ID and OTP next ID
-    for (final key in [_deviceIdPref, _otpNextIdPref]) {
-      final value = prefs.getString(key);
-      if (value != null) {
-        try {
-          await store.write(key, value);
-          if (removeFromPrefs) await prefs.remove(key);
-        } catch (e) {
-          allSucceeded = false;
-        }
-      }
-    }
-
-    if (allSucceeded) {
-      await store.write('_crypto_migration_complete', 'true');
-      debugPrint('[Crypto] Migration complete -- all keys in secure storage');
-    } else {
-      debugPrint('[Crypto] Migration incomplete -- will retry on next launch');
-    }
-  }
-
-  /// Restore identity, signing, and signed prekey from secure storage.
-  ///
-  /// [readKey] is used to read values, allowing callers to inject a fallback
-  /// strategy (e.g. SharedPreferences on web) when SecureKeyStore fails.
-  Future<void> _restoreKeysFromStorage(
-    SecureKeyStore store,
-    String storedPrivate, {
-    Future<String?> Function(String key)? readKey,
-  }) async {
-    Future<String?> read(String key) async =>
-        readKey != null ? readKey(key) : store.read(key);
-
-    final privateBytes = base64Decode(storedPrivate);
-    final publicBytes = base64Decode((await read(_identityPubKeyPref))!);
-    _identityKeyPair = SimpleKeyPairData(
-      privateBytes,
-      publicKey: SimplePublicKey(publicBytes, type: KeyPairType.x25519),
-      type: KeyPairType.x25519,
-    );
-
-    // Restore Ed25519 signing key pair
-    final sigPriv = await read(_signingKeyPref);
-    final sigPub = await read(_signingPubKeyPref);
-    if (sigPriv != null && sigPub != null) {
-      _signingKeyPair = SimpleKeyPairData(
-        base64Decode(sigPriv),
-        publicKey: SimplePublicKey(
-          base64Decode(sigPub),
-          type: KeyPairType.ed25519,
-        ),
-        type: KeyPairType.ed25519,
-      );
-    } else {
-      _signingKeyPair = await _ed25519.newKeyPair();
-      await _saveSigningKey(store);
-      _keysAreFresh = true;
-    }
-
-    // Restore signed prekey pair
-    final spkPriv = await read(_signedPrekeyPref);
-    final spkPub = await read(_signedPrekeyPubPref);
-    if (spkPriv != null && spkPub != null) {
-      _signedPrekeyPair = SimpleKeyPairData(
-        base64Decode(spkPriv),
-        publicKey: SimplePublicKey(
-          base64Decode(spkPub),
-          type: KeyPairType.x25519,
-        ),
-        type: KeyPairType.x25519,
-      );
-    } else {
-      _signedPrekeyPair = await _x25519.newKeyPair();
-      await _saveSignedPrekey(store);
-      _keysAreFresh = true;
-    }
-
-    // Always mark identity bundle for upload so the server has the
-    // current bundle.
-    _keysAreFresh = true;
-    _needsOtpReplenishment = false;
-
-    await _loadSessions(store);
-    await _rotateSignedPrekeyIfNeeded(store);
-  }
-
-  /// Initialize: load or generate identity key pair, signing key, and signed prekey.
-  Future<void> init() async {
-    try {
-      // Run migration before anything else -- moves keys from SharedPreferences
-      // into platform-secure storage if they exist there from a previous version.
-      await _migrateFromSharedPreferences();
-
-      final store = SecureKeyStore.instance;
-
-      // On web, SecureKeyStore encrypts values with Web Crypto API. If the
-      // encryption key is lost (browser cleared storage, incognito, etc.),
-      // reads return null even though the keys were previously stored. Fall
-      // back to SharedPreferences which uses plain localStorage.
-      final SharedPreferences? webFallbackPrefs = kIsWeb
-          ? await SharedPreferences.getInstance()
-          : null;
-
-      Future<String?> readWithFallback(String key) async {
-        final value = await store.read(key);
-        if (value != null) return value;
-        return webFallbackPrefs?.getString(key);
-      }
-
-      // Load or generate a unique device ID for this installation.
-      final storedDeviceId = await readWithFallback(_deviceIdPref);
-      await _loadOrGenerateDeviceId(store, storedDeviceId);
-
-      final storedPrivate = await readWithFallback(_identityKeyPref);
-
-      if (storedPrivate != null) {
-        await _restoreKeysFromStorage(
-          store,
-          storedPrivate,
-          readKey: readWithFallback,
-        );
-      } else {
-        await _generateFreshKeys(store, isFirstInstall: storedDeviceId == null);
-      }
-    } catch (e) {
-      rethrow;
-    }
-  }
-
-  /// Load an existing device ID from [storedDeviceId] or generate a new one.
-  Future<void> _loadOrGenerateDeviceId(
-    SecureKeyStore store,
-    String? storedDeviceId,
-  ) async {
-    if (storedDeviceId != null) {
-      _deviceId = int.tryParse(storedDeviceId) ?? 0;
-    } else {
-      // Generate a random positive device ID (1..2^30) to avoid collision
-      // with legacy device_id=0 from single-device era.
-      _deviceId = Random.secure().nextInt(1 << 30) + 1;
-      await store.write(_deviceIdPref, _deviceId.toString());
-      debugPrint('[Crypto] Generated new device_id: $_deviceId');
-    }
-  }
-
-  /// Generate all identity/signing/prekey key pairs for a fresh installation
-  /// (no stored private key found). Purges any stale sessions that reference
-  /// the old keys.
-  Future<void> _generateFreshKeys(
-    SecureKeyStore store, {
-    required bool isFirstInstall,
-  }) async {
-    // Only flag as "regenerated" when prior keys existed (device ID was
-    // already assigned). On a true first install there is nothing to
-    // regenerate, so suppress the misleading warning.
-    _keysWereRegenerated = !isFirstInstall;
-    _identityKeyPair = await _x25519.newKeyPair();
-    _signingKeyPair = await _ed25519.newKeyPair();
-    _signedPrekeyPair = await _x25519.newKeyPair();
-
-    final privateBytes = await (_identityKeyPair as SimpleKeyPairData)
-        .extractPrivateKeyBytes();
-    final publicKey = await _identityKeyPair!.extractPublicKey();
-
-    await store.write(_identityKeyPref, base64Encode(privateBytes));
-    await store.write(_identityPubKeyPref, base64Encode(publicKey.bytes));
-    await _saveSigningKey(store);
-    await _saveSignedPrekey(store);
-    await store.write(
-      _signedPrekeyCreatedAtPref,
-      DateTime.now().toIso8601String(),
-    );
-
-    _keysAreFresh = true;
-    _needsOtpReplenishment = true; // Fresh install needs OTP keys
-
-    // Purge any stale sessions from storage — they reference the old keys.
-    final allEntries = await store.readAll();
-    for (final key in allEntries.keys) {
-      if (key.startsWith(_sessionPrefix)) {
-        await store.delete(key);
-      }
-    }
-  }
-
-  /// Rotate the signed prekey if it is older than [_signedPrekeyMaxAge].
-  ///
-  /// The old key pair is kept as a "previous" signed prekey for a grace period
-  /// so that peers who fetched the old bundle can still complete X3DH.
-  Future<void> _rotateSignedPrekeyIfNeeded(SecureKeyStore store) async {
-    final createdAtStr = await store.read(_signedPrekeyCreatedAtPref);
-    if (createdAtStr == null) {
-      // No timestamp -- store current time and skip rotation this cycle.
-      await store.write(
-        _signedPrekeyCreatedAtPref,
-        DateTime.now().toIso8601String(),
-      );
-      return;
-    }
-
-    final createdAt = DateTime.tryParse(createdAtStr);
-    if (createdAt == null) return;
-
-    final age = DateTime.now().difference(createdAt);
-    if (age < _signedPrekeyMaxAge) return;
-
-    debugPrint('[Crypto] Signed prekey is ${age.inDays} days old -- rotating');
-
-    // Move current signed prekey to "previous" slot
-    final currentPriv = await store.read(_signedPrekeyPref);
-    final currentPub = await store.read(_signedPrekeyPubPref);
-    if (currentPriv != null && currentPub != null) {
-      await store.write(_signedPrekeyPreviousPref, currentPriv);
-      await store.write(_signedPrekeyPreviousPubPref, currentPub);
-    }
-
-    // Generate new signed prekey
-    _signedPrekeyPair = await _x25519.newKeyPair();
-    await _saveSignedPrekey(store);
-    await store.write(
-      _signedPrekeyCreatedAtPref,
-      DateTime.now().toIso8601String(),
-    );
-
-    _keysAreFresh = true;
-
-    // Clean up previous prekey if it has exceeded the grace period
-    await _cleanupPreviousPrekey(store);
-  }
-
-  /// Remove the previous signed prekey if it is older than the grace period.
-  Future<void> _cleanupPreviousPrekey(SecureKeyStore store) async {
-    final prevPriv = await store.read(_signedPrekeyPreviousPref);
-    if (prevPriv == null) return;
-
-    // The previous prekey was created when the current one replaced it.
-    // We use the current prekey creation time minus max age as an estimate
-    // for when the previous key was active. For simplicity, keep it for
-    // one full grace period from now -- it will be cleaned up on the next
-    // rotation cycle after that.
-    final createdAtStr = await store.read(_signedPrekeyCreatedAtPref);
-    if (createdAtStr == null) return;
-
-    final createdAt = DateTime.tryParse(createdAtStr);
-    if (createdAt == null) return;
-
-    // If the current prekey is already older than the grace period minus
-    // the max age, the previous one has definitely expired.
-    final currentAge = DateTime.now().difference(createdAt);
-    if (currentAge >= _signedPrekeyGracePeriod) {
-      await store.delete(_signedPrekeyPreviousPref);
-      await store.delete(_signedPrekeyPreviousPubPref);
-      debugPrint('[Crypto] Cleaned up expired previous signed prekey');
-    }
-  }
-
-  /// Fetch and cache a peer's identity public key from the server.
-  ///
-  /// Returns the key bytes, or null if unavailable.
-  /// Uses Trust-On-First-Use (TOFU): the first-seen key is trusted and
-  /// persisted. If a later fetch returns a different key, a warning is logged
-  /// and a `_peerIdentityChangedPrefix` flag is persisted for the peer.
-  Future<Uint8List?> fetchPeerIdentityKey(String peerUserId) async {
-    // Check cache first
-    final store = SecureKeyStore.instance;
-    final cached = await store.read('$_peerIdentityPrefix$peerUserId');
-    if (cached != null) return Uint8List.fromList(base64Decode(cached));
-
-    // Fetch from server
-    try {
-      final response = await http.get(
-        Uri.parse('$serverUrl/api/keys/bundle/$peerUserId'),
-        headers: {'Authorization': 'Bearer $_token'},
-      );
-      if (response.statusCode != 200) return null;
-
-      final data = jsonDecode(response.body) as Map<String, dynamic>;
-      final identityKeyB64 = data['identity_key'] as String?;
-      if (identityKeyB64 == null) return null;
-
-      await _storePeerIdentityKeyTofu(peerUserId, identityKeyB64);
-      return Uint8List.fromList(base64Decode(identityKeyB64));
-    } catch (e) {
-      debugPrint('[Crypto] Failed to fetch peer identity key: $e');
-      return null;
-    }
-  }
-
-  /// Store a peer's identity key with Trust-On-First-Use (TOFU) semantics.
-  ///
-  /// - First encounter: stores the key.
-  /// - Same key: no-op.
-  /// - Different key: logs a warning via [DebugLogService] and persists a
-  ///   flag (`_peerIdentityChangedPrefix + peerId`) for UI to surface later.
-  ///   The new key is stored so future sessions use the latest key.
-  Future<void> _storePeerIdentityKeyTofu(
-    String peerUserId,
-    String newKeyB64,
-  ) async {
-    final store = SecureKeyStore.instance;
-    final storeKey = '$_peerIdentityPrefix$peerUserId';
-    final existing = await store.read(storeKey);
-
-    if (existing != null && existing != newKeyB64) {
-      DebugLogService.instance.log(
-        LogLevel.warning,
-        'Crypto',
-        'TOFU: identity key changed for peer $peerUserId -- '
-            'possible key reset or MITM. Old prefix: '
-            '${existing.substring(0, min(8, existing.length))}..., '
-            'new prefix: '
-            '${newKeyB64.substring(0, min(8, newKeyB64.length))}...',
-      );
-      debugPrint('[Crypto] WARNING: peer $peerUserId identity key changed!');
-      // Persist a flag so the UI can surface a warning banner later.
-      await store.write(
-        '$_peerIdentityChangedPrefix$peerUserId',
-        DateTime.now().toIso8601String(),
-      );
-    }
-
-    // Store the (possibly new) key so future sessions use the latest.
-    await store.write(storeKey, newKeyB64);
-  }
-
-  /// Check whether a peer's identity key has changed since first contact.
-  Future<bool> hasPeerIdentityKeyChanged(String peerUserId) async {
-    final store = SecureKeyStore.instance;
-    final flag = await store.read('$_peerIdentityChangedPrefix$peerUserId');
-    return flag != null;
-  }
-
-  /// Acknowledge a peer identity key change (clears the flag).
-  Future<void> acknowledgePeerIdentityKeyChange(String peerUserId) async {
-    final store = SecureKeyStore.instance;
-    await store.delete('$_peerIdentityChangedPrefix$peerUserId');
-  }
-
-  /// Explicitly trust the peer's new identity key after the user has
-  /// reviewed (or chosen to skip reviewing) the safety number.
-  ///
-  /// This is the explicit counterpart to the previously-silent overwrite
-  /// in TOFU: it clears the change flag, persists [newIdentityKeyB64] (or,
-  /// if `null`, leaves whatever was last written by TOFU in place), and
-  /// drops any cached/persisted Signal session for the peer so the next
-  /// outbound message triggers a fresh X3DH against the trusted key.
-  /// (#580)
-  Future<void> acceptIdentityKeyChange(
-    String peerUserId, {
-    String? newIdentityKeyB64,
-  }) async {
-    final store = SecureKeyStore.instance;
-    if (newIdentityKeyB64 != null) {
-      await store.write('$_peerIdentityPrefix$peerUserId', newIdentityKeyB64);
-    }
-    await store.delete('$_peerIdentityChangedPrefix$peerUserId');
-    // Drop any cached/persisted session keyed to the old identity so the
-    // next send re-runs X3DH against the freshly-trusted key.
-    _sessions.remove(peerUserId);
-    await store.delete('$_sessionPrefix$peerUserId');
-  }
-
-  /// Compute the safety-number fingerprint between this device and
-  /// [peerUserId], or `null` if either identity key is unavailable.
-  ///
-  /// Deterministic regardless of which side calls it (keys are sorted
-  /// before hashing). See [SafetyNumberService] for the spec.
-  Future<String?> safetyNumberFor(String peerUserId) async {
-    final myKey = await getIdentityPublicKey();
-    if (myKey == null) return null;
-    final store = SecureKeyStore.instance;
-    final peerB64 = await store.read('$_peerIdentityPrefix$peerUserId');
-    if (peerB64 == null) return null;
-    final peerKey = Uint8List.fromList(base64Decode(peerB64));
-    return SafetyNumberService.generate(myKey, peerKey);
-  }
-
-  /// Get the local identity public key bytes.
-  Future<Uint8List?> getIdentityPublicKey() async {
-    if (_identityKeyPair == null) return null;
-    final pub = await _identityKeyPair!.extractPublicKey();
-    return Uint8List.fromList(pub.bytes);
-  }
+  // Peer identity / TOFU / safety-number helpers live in
+  // crypto/peer_identity_extension.dart (extension CryptoServicePeerIdentity).
 
   Future<void> _saveSigningKey(SecureKeyStore store) async {
     final sigPrivBytes = await (_signingKeyPair as SimpleKeyPairData)
@@ -870,105 +355,8 @@ class CryptoService {
     }
   }
 
-  /// Generate 10 new OTP key pairs with monotonically increasing IDs.
-  ///
-  /// Uses a persisted counter ([_otpNextIdPref]) so that IDs never collide
-  /// with previously uploaded keys. This prevents the critical bug where
-  /// re-using IDs 0-9 on every upload causes the server to keep old public
-  /// keys (via ON CONFLICT DO NOTHING) while the client overwrites local
-  /// private keys with new material.
-  Future<void> _generateAndPersistOtpKeys(
-    List<Map<String, dynamic>> otps,
-  ) async {
-    final store = SecureKeyStore.instance;
-
-    // Read persisted counter (default 0 for first-ever upload)
-    int nextId = 0;
-    final storedNextId = await store.read(_otpNextIdPref);
-    if (storedNextId != null) {
-      nextId = int.tryParse(storedNextId) ?? 0;
-    }
-
-    for (var i = 0; i < 10; i++) {
-      final keyId = nextId + i;
-      final otpPair = await _x25519.newKeyPair();
-      final otpPub = await otpPair.extractPublicKey();
-      final otpPrivBytes = await (otpPair as SimpleKeyPairData)
-          .extractPrivateKeyBytes();
-
-      final pubB64 = base64Encode(otpPub.bytes);
-      await store.write(
-        '$_otpPrivatePrefix$keyId',
-        jsonEncode({'private': base64Encode(otpPrivBytes), 'public': pubB64}),
-      );
-
-      otps.add({'key_id': keyId, 'public_key': pubB64});
-    }
-
-    // Persist the next counter value for future uploads
-    await store.write(_otpNextIdPref, '${nextId + 10}');
-    debugPrint('[Crypto] Generated OTP keys $nextId..${nextId + 9}');
-  }
-
-  /// Check the server's remaining unused OTP count and replenish if low.
-  Future<void> checkAndReplenishOtpKeys() async {
-    try {
-      final response = await http.get(
-        Uri.parse('$serverUrl/api/keys/otp-count?device_id=$_deviceId'),
-        headers: {'Authorization': 'Bearer $_token'},
-      );
-      if (response.statusCode == 200) {
-        final data = jsonDecode(response.body) as Map<String, dynamic>;
-        final count = data['count'] as int? ?? 0;
-        if (count < 5) {
-          debugPrint('[Crypto] OTP count low ($count), replenishing...');
-          final otps = <Map<String, dynamic>>[];
-          await _generateAndPersistOtpKeys(otps);
-          // Upload just the new OTPs (identity bundle already on server)
-          await _uploadOtpBatch(otps);
-        }
-      }
-    } catch (e) {
-      debugPrint('[Crypto] OTP replenishment check failed: $e');
-    }
-  }
-
-  /// Upload a batch of OTP public keys to the server.
-  Future<void> _uploadOtpBatch(List<Map<String, dynamic>> otps) async {
-    if (otps.isEmpty) return;
-    if (_identityKeyPair == null) return;
-
-    final publicKey = await _identityKeyPair!.extractPublicKey();
-    final spkPub = await _signedPrekeyPair!.extractPublicKey();
-    final signature = await _ed25519.sign(
-      spkPub.bytes,
-      keyPair: _signingKeyPair!,
-    );
-    final signingPub = await _signingKeyPair!.extractPublicKey();
-
-    final body = jsonEncode({
-      'identity_key': base64Encode(publicKey.bytes),
-      'signing_key': base64Encode(signingPub.bytes),
-      'signed_prekey': base64Encode(spkPub.bytes),
-      'signed_prekey_signature': base64Encode(signature.bytes),
-      'signed_prekey_id': 1,
-      'one_time_prekeys': otps,
-      'device_id': _deviceId,
-    });
-
-    final response = await http.post(
-      Uri.parse('$serverUrl/api/keys/upload'),
-      headers: {
-        _contentTypeHeader: _applicationJson,
-        'Authorization': 'Bearer $_token',
-      },
-      body: body,
-    );
-
-    if (response.statusCode == 201) {
-      debugPrint('[Crypto] Uploaded ${otps.length} new OTP keys');
-    }
-  }
+  // OTP generation / replenishment helpers live in
+  // crypto/otp_extension.dart (extension CryptoServiceOtp).
 
   /// Get or create a Signal session with a peer.
   ///
